lalc_backend/pyproject.toml
[project]
name = "lalcbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=1.26.4",
    "onnxruntime>=1.23.2",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "plyer>=2.1.0",
    "psutil>=7.1.3",
    "pynput>=1.8.1",
    "pywin32>=311; platform_system == 'Windows'",
    "pyobjc-framework-Quartz>=10.0; platform_system == 'Darwin'",
    "rapidocr>=3.4.3",
    "scipy>=1.15.3",
    "websockets>=15.0.1",
]

[tool.setuptools.packages.find]
include = ["workflow*", "input*", "utils*", "recognize*"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
